NSecureFacePy/requirements.txt
boto3
Flask
python-dotenv==0.10.3
opencv-python
numpy
imutils
dlib
asyncio==3.4.3
aiohttp=3.5.4
flask-wtf